+++ waterfall-cad/CHANGELOG.md
@@ -8,10 +8,12 @@ and this project adheres to the
 
 ## Unreleased
 
-## 0.6.1.0
+## 0.6.2.0
 
 - Added `Waterfall.Path.Common.pathLength`, `Waterfall.Path.pathLength3D` and `Waterfall.TwoD.Path2D.pathLength2D`
 - Added `Waterfall.Path.Common.takePathFraction`, `Waterfall.Path.takePathFraction3D` and `Waterfall.TwoD.Path2D.takePathFraction2D`
+
+## 0.6.1.0
 
 ### Added